Refactor the routing endpoints
