+++ front-end/components/BasicSection.js
@@ -23,8 +23,6 @@ const BasicSection = ({ handleInputChange, setFormData}) => {
     const [projectTitle, setProjectTitle] = React.useState('');
     const [selectedFields, setSelectedFields] = React.useState([]);
     const [abstract, setAbstract] = React.useState('');
-    // const [budget, setBudget] = React.useState('');
-    // const [objectives, setObjectives] = React.useState('');
 
     const handleFieldSelectionChange = (event) => {
         const { checked, name } = event.target ?? {};
@@ -101,36 +99,7 @@ const BasicSection = ({ handleInputChange, setFormData}) => {
                     inputProps={{ maxLength: 2000 }}
                     name="abstract"
                 />
-
-                {/*<TextField*/}
-                {/*    label="Budget"*/}
-                {/*    value={budget}*/}
-                {/*    onChange={(event) => {*/}
-                {/*        setBudget(event.target.value);*/}
-                {/*        handleInputChange(event);*/}
-                {/*    }}*/}
-                {/*    variant="outlined"*/}
-                {/*    // multiline*/}
-                {/*    // maxRows={10}*/}
-                {/*    inputProps={{ maxLength: 50}}*/}
-                {/*    name="budget"*/}
-                {/*/>*/}
-
-                {/*<TextField*/}
-                {/*    label="Project Objectives"*/}
-                {/*    value={objectives}*/}
-                {/*    onChange={(event) => {*/}
-                {/*        setObjectives(event.target.value);*/}
-                {/*        handleInputChange(event);*/}
-                {/*    }}*/}
-                {/*    variant="outlined"*/}
-                {/*    multiline*/}
-                {/*    maxRows={50}*/}
-                {/*    inputProps={{ maxLength: 2000}}*/}
-                {/*    name="objectives"*/}
-                {/*/>*/}
             </Box>
-
         </Box>
     );
 };
