+++ front-end/pages/admin/panel.js
