+++ front-end/pages/researcher/project_submit.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import BasicSection from "../components/BasicSection";
-import BudgetSection from "../components/BudgetSection";
-import TimelineSection from "../components/TimelineSection";
+import BasicSection from "../../components/BasicSection";
+import BudgetSection from "../../components/BudgetSection";
+import TimelineSection from "../../components/TimelineSection";
 import {Box, Button} from "@mui/material";
 
 function ProjectSubmit() {
@@ -34,6 +34,7 @@ function ProjectSubmit() {
         })
             .then(response => {
                 console.log(response);
+                // console.log('Role:', role);
                 // Handle response from the API
             })
             .catch(error => {
